feat: Extend deploy script to install for all kernels
If, for whatever reason, this Kernel module is installed freshly, the
existing deploy script would not install the module for all available
Kernels. If a user would boot into another Kernel (e.g. 6.12.21-amd64
instead of 6.12.22-amd64) Wi-Fi would not be available. This can be
confusing and annoying.

The changes of this commit would prune the module from DKMS and then
install it for all available Kernels. This is much more convenient.